Remove any cyclic packages
