+++ app/src/lib/ai/anthropicClient.ts
@@ -1,6 +1,10 @@
-import { type AIClient, type PromptMessage, MessageRole } from '$lib/backend/aiClient';
+import {
+	MessageRole,
+	type AIClient,
+	type AnthropicModelName,
+	type PromptMessage
+} from '$lib/aiTypes';
 import { fetch, Body } from '@tauri-apps/api/http';
-import type { AnthropicModelName } from './types';
 
 type AnthropicAPIResponse = { content: { text: string }[] };
